requirements.txt
aiohttp==3.9.5
fastapi==0.111.1
fastapi_cache==0.1.0
httpx==0.27.0
motor==3.1.2
pydantic==1.10.7
pymongo==4.8.0
python_bcrypt==0.3.2
redis==4.6.0
starlette==0.38.1
uvicorn==0.30.3
